Add zone sizing (#15)
* Add next zone choosing

* Add unit test for zone tick

* Fix zone tick unit tests

* Add start zone tick

* Add simple start unit test

* Fix tick start unit test

* Add sub zone event api method

* Add zone controller check

* Add start zone controller on spot start

* Add sub

* Fix server zone ticker

* Fix log and remove unused fields

* Fix zone tick bugs

* Fix zone position bug

* Improve accuracy

* Fix zone tick algorithm

* Fix zone ticking start time

* Fix get initial zone

* Fix game UI

* Fix show error on spot settings page

* Add feature for chaning env in realtime

* Fix client tests

* Fix unit tests and lint

+++ server/pkg/player_state/player_state.go
@@ -12,7 +12,7 @@ import (
 
 type PlayerState struct {
 	Position s2.LatLng
-	Health   int16
+	Health   float32
 
 	// Zone damage
 
@@ -74,5 +74,5 @@ func (m *PlayerStateMap) Store(key uuid.UUID, value PlayerState) {
 type PlayerPublicState struct {
 	PlayerUuid uuid.UUID
 	Position   s2.LatLng
-	Health     int16
+	Health     float32
 }
